LYNX-738: Add Catalog Price Rules Support for Targeted Blocks (#305)
